+++ Dev/Cpp/EffekseerSoundOSMixer/EffekseerSound/EffekseerSoundOSMixer.SoundPlayer.cpp
@@ -75,7 +75,7 @@ SoundPlayer::~SoundPlayer()
 	instance.data = soundData;
 	m_sound->AddInstance(instance);
 
-	return reinterpret_cast<Effekseer::SoundHandle>(id);
+	return reinterpret_cast<Effekseer::SoundHandle>(static_cast<size_t>(id));
 }
 
 //----------------------------------------------------------------------------------
